Project(WHMS)
I'm developing a standalone application for the ‘Postlime private Ltd’. The company manufactures cement & has warehouses around the country to store supplies for the cement factory near them. I'm going to develop a partial ‘Warehouse Management System’ for the company. We will only consider specific interactions of the ‘Area Manager’ & the ‘Warehouse Supervisor’ with the application.